[P2P-151] refactor: 파일 위치 수정(config 파일끼리 묶음)
